# openapi.yaml
openapi: 3.1.0
info:
  title: Shoutboard API
  version: 1.0.0
  description: |
    API-first employee recognition group card platform. Shoutboard enables teams to
    create collaborative recognition cards for birthdays, promotions, farewells, and
    other occasions with public shareable links and team contributions.

    ## Key Features
    - **Multi-occasion support**: Birthday, anniversary, promotion, welcome, farewell, custom
    - **Public sharing**: Generate shareable links for public contribution
    - **Team collaboration**: Collect messages from multiple contributors anonymously or identified
    - **Webhook integrations**: Real-time event notifications for board creation, posting, and delivery
    - **API-first design**: Full programmatic access with JWT or API key authentication
    - **Analytics**: Track board engagement, contributor metrics, and team participation

    ## Authentication Methods
    The API supports two authentication methods:

    ### JWT Bearer Token
    Register and login to receive a JWT token valid for 7 days. Include in Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ### API Key
    Create persistent API keys in the dashboard for service-to-service integrations.
    Include in header:
    ```
    X-API-Key: sb_live_<random_string>
    ```
    API keys can be scoped to specific permissions and revoked at any time.

    ## Webhook Events
    Subscribe to real-time events:
...

...

    ## Rate Limiting
...

...
    ```json
...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: string
        authorName:
          type: string
        contentText:
          type: string
          maxLength: 5000
        mediaUrl:
          type: string
          format: uri
          nullable: true
        gifUrl:
          type: string
          format: uri
          nullable: true
        isAnonymous:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, authorName, contentText, isAnonymous, createdAt]
...
    Webhook:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum: [board.created, board.sent, post.created, post.deleted, '*']
          description: Events to subscribe to, or '*' for all events
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        secret:
          type: string
          description: Signing secret (only returned on creation)
      required: [id, url, events, active, createdAt]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]
...
paths:
...
  /v1/boards/{boardId}/posts:
    get:
      summary: List posts for a board
      operationId: listPosts
      tags: [Posts]
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      description: Public endpoint for SENT boards, authenticated for DRAFT/ACTIVE.
      responses:
        200:
          description: List of posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      posts:
                        type: array
                        items:
                          $ref: '#/components/schemas/Post'
                      pagination:
                        type: object
                        properties:
                          limit:
                            type: integer
                          offset:
                            type: integer
                          total:
                            type: integer

    post:
      summary: Add post to board
      operationId: createPost
      tags: [Posts]
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
      description: |
        Public for SENT boards (no auth required), authenticated for DRAFT/ACTIVE boards.
        Contributors can submit messages anonymously or identified.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                authorName:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: John Smith
                contentText:
                  type: string
                  minLength: 1
                  maxLength: 5000
                  example: "Happy birthday! Wishing you an amazing year ahead!"
                mediaUrl:
                  type: string
                  format: uri
                  nullable: true
                gifUrl:
                  type: string
                  format: uri
                  nullable: true
                isAnonymous:
                  type: boolean
                  default: false
              required: [authorName, contentText]
      responses:
        201:
          description: Post created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Post'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'
        422:
          $ref: '#/components/responses/ValidationError'
  /v1/boards/{boardId}/posts/{postId}:
    delete:
      summary: Delete a post
      operationId: deletePost
      tags: [Posts]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      description: Only board creator or admin can delete posts.
      responses:
        200:
          description: Post deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'

  # API Key Endpoints
...
  /v1/webhooks:
    get:
      summary: List webhook subscriptions (admin only)
      operationId: listWebhooks
      tags: [Webhooks]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      webhooks:
                        type: array
                        items:
                          $ref: '#/components/schemas/Webhook'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'

    post:
      summary: Create webhook subscription (admin only)
      operationId: createWebhook
      tags: [Webhooks]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                  example: https://example.com/webhooks/shoutboard
                events:
                  type: array
                  items:
                    type: string
                    enum: [board.created, board.sent, post.created, post.deleted, '*']
                  minItems: 1
                  example: [board.created, board.sent, post.created]
              required: [url, events]
      responses:
        201:
          description: Webhook created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Webhook'
                      - type: object
                        properties:
                          secret:
                            type: string
                            description: Signing secret - store securely for verifying signatures
        403:
          $ref: '#/components/responses/ForbiddenError'
        422:
          $ref: '#/components/responses/ValidationError'
  /v1/webhooks/{id}:
    patch:
      summary: Update webhook configuration (admin only)
      operationId: updateWebhook
      tags: [Webhooks]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                active:
                  type: boolean
      responses:
        200:
          description: Webhook updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Webhook'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete webhook subscription (admin only)
      operationId: deleteWebhook
      tags: [Webhooks]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Webhook deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'
...
components:
  responses:
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Board not found
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Insufficient permissions
    ValidationError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Field validation failed